+++ source/Relativity.Testing.Framework.Api.Tests/Strategies/Imaging/Set/ImagingSetHideStrategyNotSupportedFixture.cs
@@ -6,35 +6,35 @@ using Relativity.Testing.Framework.Api.Strategies;
 namespace Relativity.Testing.Framework.Api.Tests.Strategies
 {
 	[TestFixture]
-	[TestOf(typeof(ImagingSetCreateStrategyNotSupported))]
-	public class ImagingSetGetStrategyNotSupportedFixture
+	[TestOf(typeof(ImagingSetHideStrategyNotSupported))]
+	public class ImagingSetHideStrategyNotSupportedFixture
 	{
-		private const string _NOT_SUPPORTED_EXCEPTION_MESSAGE = "The method Get Imaging Set does not support version of Relativity lower than 12.1.";
+		private const string _NOT_SUPPORTED_EXCEPTION_MESSAGE = "The method Hide Imaging Set does not support version of Relativity lower than 12.1.";
 		private const int _WORKSPACE_ID = 10000;
 		private const int _IMAGING_SET_ID = 10001;
 
-		private ImagingSetGetStrategyNotSupported _sut;
+		private ImagingSetHideStrategyNotSupported _sut;
 
 		[SetUp]
 		public void SetUp()
 		{
-			_sut = new ImagingSetGetStrategyNotSupported();
+			_sut = new ImagingSetHideStrategyNotSupported();
 		}
 
 		[Test]
-		public void Get_ForNotSupportedVersion_ThrowsArgumentException()
+		public void Hide_ForNotSupportedVersion_ThrowsArgumentException()
 		{
 			var result = Assert.Throws<ArgumentException>(() =>
-				_sut.Get(_WORKSPACE_ID, _IMAGING_SET_ID));
+				_sut.Hide(_WORKSPACE_ID, _IMAGING_SET_ID));
 
 			result.Message.Should().Contain(_NOT_SUPPORTED_EXCEPTION_MESSAGE);
 		}
 
 		[Test]
-		public void GetAsync_ForNotSupportedVersion_ThrowsArgumentException()
+		public void HideAsync_ForNotSupportedVersion_ThrowsArgumentException()
 		{
 			var result = Assert.ThrowsAsync<ArgumentException>(async () =>
-				await _sut.GetAsync(_WORKSPACE_ID, _IMAGING_SET_ID).ConfigureAwait(false));
+				await _sut.HideAsync(_WORKSPACE_ID, _IMAGING_SET_ID).ConfigureAwait(false));
 
 			result.Message.Should().Contain(_NOT_SUPPORTED_EXCEPTION_MESSAGE);
 		}