+++ source/Relativity.Testing.Framework.Api/Strategies/Imaging/Job/DTO/ImagingJobIdResponseDtoV1.cs
@@ -1,0 +1,7 @@
+﻿namespace Relativity.Testing.Framework.Api.Strategies
+{
+	internal class ImagingJobIdResponseDtoV1
+	{
+		public int ImagingJobID { get; set; }
+	}
+}