+++ source/Relativity.Testing.Framework.Api/Strategies/Imaging/Set/V1/ImagingSetCreateStrategyV1.cs
@@ -8,7 +8,7 @@ namespace Relativity.Testing.Framework.Api.Strategies
 	[VersionRange(">=12.1")]
 	internal class ImagingSetCreateStrategyV1 : IImagingSetCreateStrategy
 	{
-		private const string _BASE_GET_URL_FORMAT = "relativity-imaging/v1/workspaces/{0}/imaging-sets";
+		private const string _BASE_CREATE_URL_FORMAT = "relativity-imaging/v1/workspaces/{0}/imaging-sets";
 		private readonly IRestService _restService;
 		private readonly IImagingSetValidatorV1 _imagingSetValidator;
 		private readonly IImagingSetGetStrategy _imagingSetGetStrategy;
@@ -25,10 +25,10 @@ namespace Relativity.Testing.Framework.Api.Strategies
 
 		public ImagingSet Create(int workspaceId, ImagingSetRequest imagingSetRequest)
 		{
-			_imagingSetValidator.ValidateImagingRequest(workspaceId, imagingSetRequest);
+			_imagingSetValidator.ValidateImagingSetCreateRequest(workspaceId, imagingSetRequest);
 
 			string url = GetUrlWithParameters(workspaceId);
-			var dto = new ImagingSetRequestDTOV1(imagingSetRequest);
+			var dto = new ImagingSetRequestDtoV1(imagingSetRequest);
 
 			var createdImagingSetId = _restService.Post<int>(url, dto);
 
@@ -37,10 +37,10 @@ namespace Relativity.Testing.Framework.Api.Strategies
 
 		public async Task<ImagingSet> CreateAsync(int workspaceId, ImagingSetRequest imagingSetRequest)
 		{
-			_imagingSetValidator.ValidateImagingRequest(workspaceId, imagingSetRequest);
+			_imagingSetValidator.ValidateImagingSetCreateRequest(workspaceId, imagingSetRequest);
 
 			string url = GetUrlWithParameters(workspaceId);
-			var dto = new ImagingSetRequestDTOV1(imagingSetRequest);
+			var dto = new ImagingSetRequestDtoV1(imagingSetRequest);
 
 			var createdImagingSetId = await _restService.PostAsync<int>(url, dto).ConfigureAwait(false);
 
@@ -49,7 +49,7 @@ namespace Relativity.Testing.Framework.Api.Strategies
 
 		private static string GetUrlWithParameters(int workspaceId)
 		{
-			return string.Format(_BASE_GET_URL_FORMAT, workspaceId);
+			return string.Format(_BASE_CREATE_URL_FORMAT, workspaceId);
 		}
 	}
 }